pyproject.toml
[tool.poetry]
name = "notification-server"
version = "0.1.0"
description = ""
authors = ["Md Reaz Ahammed Chowdhury <reaz.ahammed@welldev.io>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
pika = "^1.3.2"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
pydantic = "^2.9.2"
pydantic-settings = "^2.6.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
